Implementation of DigitBoard class
implemented
